Combine tests into one test
